+++ opm/simulators/utils/readDeck.cpp
@@ -1,0 +1,255 @@
+/*
+  Copyright 2013, 2014, 2015 SINTEF ICT, Applied Mathematics.
+  Copyright 2014 Dr. Blatt - HPC-Simulation-Software & Services
+  Copyright 2015 IRIS AS
+  Copyright 2014 STATOIL ASA.
+
+  This file is part of the Open Porous Media project (OPM).
+
+  OPM is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  OPM is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+*/
+#include "config.h"
+
+#if HAVE_MPI
+#include "mpi.h"
+#endif
+#include "readDeck.hpp"
+
+#include <opm/common/utility/String.hpp>
+
+#include <opm/io/eclipse/EclIOdata.hpp>
+
+#include <opm/output/eclipse/RestartIO.hpp>
+#include <opm/io/eclipse/rst/state.hpp>
+
+#include <opm/parser/eclipse/EclipseState/checkDeck.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/ArrayDimChecker.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Schedule.hpp>
+#include <opm/parser/eclipse/Parser/ErrorGuard.hpp>
+
+#include <opm/simulators/flow/MissingFeatures.hpp>
+#include <opm/simulators/utils/ParallelEclipseState.hpp>
+#include <opm/simulators/utils/ParallelSerialization.hpp>
+
+namespace Opm
+{
+
+void ensureOutputDirExists_(const std::string& cmdline_output_dir)
+{
+    if (!Opm::filesystem::is_directory(cmdline_output_dir)) {
+        try {
+            Opm::filesystem::create_directories(cmdline_output_dir);
+        }
+        catch (...) {
+            throw std::runtime_error("Creation of output directory '" + cmdline_output_dir + "' failed\n");
+        }
+    }
+}
+
+// Setup the OpmLog backends
+FileOutputMode setupLogging(int mpi_rank_, const std::string& deck_filename, const std::string& cmdline_output_dir, const std::string& cmdline_output, bool output_cout_, const std::string& stdout_log_id) {
+
+    if (!cmdline_output_dir.empty()) {
+        ensureOutputDirExists_(cmdline_output_dir);
+    }
+
+    // create logFile
+    using Opm::filesystem::path;
+    path fpath(deck_filename);
+    std::string baseName;
+    std::ostringstream debugFileStream;
+    std::ostringstream logFileStream;
+
+    // Strip extension "." or ".DATA"
+    std::string extension = uppercase(fpath.extension().string());
+    if (extension == ".DATA" || extension == ".") {
+        baseName = uppercase(fpath.stem().string());
+    } else {
+        baseName = uppercase(fpath.filename().string());
+    }
+
+    std::string output_dir = cmdline_output_dir;
+    if (output_dir.empty()) {
+        output_dir = fpath.has_parent_path()
+            ? absolute(fpath.parent_path()).generic_string()
+            : Opm::filesystem::current_path().generic_string();
+    }
+
+    logFileStream << output_dir << "/" << baseName;
+    debugFileStream << output_dir << "/" << baseName;
+
+    if (mpi_rank_ != 0) {
+        // Added rank to log file for non-zero ranks.
+        // This prevents message loss.
+        debugFileStream << "." << mpi_rank_;
+        // If the following file appears then there is a bug.
+        logFileStream << "." << mpi_rank_;
+    }
+    logFileStream << ".PRT";
+    debugFileStream << ".DBG";
+
+    FileOutputMode output;
+    {
+        static std::map<std::string, FileOutputMode> stringToOutputMode =
+            { {"none", FileOutputMode::OUTPUT_NONE },
+              {"false", FileOutputMode::OUTPUT_LOG_ONLY },
+              {"log", FileOutputMode::OUTPUT_LOG_ONLY },
+              {"all" , FileOutputMode::OUTPUT_ALL },
+              {"true" , FileOutputMode::OUTPUT_ALL }};
+        auto outputModeIt = stringToOutputMode.find(cmdline_output);
+        if (outputModeIt != stringToOutputMode.end()) {
+            output = outputModeIt->second;
+        }
+        else {
+            output = FileOutputMode::OUTPUT_ALL;
+            std::cerr << "Value " << cmdline_output <<
+                " is not a recognized output mode. Using \"all\" instead."
+                      << std::endl;
+        }
+    }
+
+    if (output > FileOutputMode::OUTPUT_NONE) {
+        std::shared_ptr<Opm::EclipsePRTLog> prtLog = std::make_shared<Opm::EclipsePRTLog>(logFileStream.str(), Opm::Log::NoDebugMessageTypes, false, output_cout_);
+        Opm::OpmLog::addBackend("ECLIPSEPRTLOG", prtLog);
+        prtLog->setMessageLimiter(std::make_shared<Opm::MessageLimiter>());
+        prtLog->setMessageFormatter(std::make_shared<Opm::SimpleMessageFormatter>(false));
+    }
+
+    if (output >= FileOutputMode::OUTPUT_LOG_ONLY) {
+        std::string debugFile = debugFileStream.str();
+        std::shared_ptr<Opm::StreamLog> debugLog = std::make_shared<Opm::EclipsePRTLog>(debugFileStream.str(), Opm::Log::DefaultMessageTypes, false, output_cout_);
+        Opm::OpmLog::addBackend("DEBUGLOG", debugLog);
+    }
+
+    if (mpi_rank_ == 0) {
+        std::shared_ptr<Opm::StreamLog> streamLog = std::make_shared<Opm::StreamLog>(std::cout, Opm::Log::StdoutMessageTypes);
+        Opm::OpmLog::addBackend(stdout_log_id, streamLog);
+        streamLog->setMessageFormatter(std::make_shared<Opm::SimpleMessageFormatter>(true));
+    }
+    return output;
+}
+
+
+void setupMessageLimiter(const Opm::MessageLimits msgLimits,  const std::string& stdout_log_id) {
+    std::shared_ptr<Opm::StreamLog> stream_log = Opm::OpmLog::getBackend<Opm::StreamLog>(stdout_log_id);
+
+    const std::map<int64_t, int> limits = {{Opm::Log::MessageType::Note,
+                                            msgLimits.getCommentPrintLimit(0)},
+                                           {Opm::Log::MessageType::Info,
+                                            msgLimits.getMessagePrintLimit(0)},
+                                           {Opm::Log::MessageType::Warning,
+                                            msgLimits.getWarningPrintLimit(0)},
+                                           {Opm::Log::MessageType::Error,
+                                            msgLimits.getErrorPrintLimit(0)},
+                                           {Opm::Log::MessageType::Problem,
+                                            msgLimits.getProblemPrintLimit(0)},
+                                           {Opm::Log::MessageType::Bug,
+                                            msgLimits.getBugPrintLimit(0)}};
+    stream_log->setMessageLimiter(std::make_shared<Opm::MessageLimiter>(10, limits));
+}
+
+
+void readDeck(int rank, std::string& deckFilename, std::unique_ptr<Opm::Deck>& deck, std::unique_ptr<Opm::EclipseState>& eclipseState,
+              std::unique_ptr<Opm::Schedule>& schedule, std::unique_ptr<Opm::SummaryConfig>& summaryConfig,
+              std::shared_ptr<Opm::Python>& python, const ParseContext& parseContext,
+              bool initFromRestart, bool checkDeck)
+{
+    Opm::Parser parser;
+    Opm::ErrorGuard errorGuard;
+
+#if HAVE_MPI
+    int parseSuccess = 0;
+#endif
+    std::string failureMessage;
+
+    if (rank==0) {
+        try
+        {
+            if (!deck)
+                deck.reset( new Opm::Deck( parser.parseFile(deckFilename , parseContext, errorGuard)));
+            Opm::MissingFeatures::checkKeywords(*deck, parseContext, errorGuard);
+            if ( checkDeck )
+                Opm::checkDeck(*deck, parser, parseContext, errorGuard);
+
+            if (!eclipseState) {
+#if HAVEMPI
+                eclipseState.reset(new Opm::ParallelEclipseState(*deck));
+#else
+                eclipseState.reset(new Opm::EclipseState(*deck));
+#endif
+            }
+            /*
+              For the time being initializing wells and groups from the
+              restart file is not possible, but work is underways and it is
+              included here as a switch.
+            */
+            const auto& init_config = eclipseState->getInitConfig();
+            if (init_config.restartRequested() && initFromRestart) {
+                int report_step = init_config.getRestartStep();
+                const auto& rst_filename = eclipseState->getIOConfig().getRestartFileName( init_config.getRestartRootName(), report_step, false );
+                Opm::EclIO::ERst rst_file(rst_filename);
+                const auto& rst_state = Opm::RestartIO::RstState::load(rst_file, report_step);
+                if (!schedule)
+                    schedule.reset(new Opm::Schedule(*deck, *eclipseState, parseContext, errorGuard, python, &rst_state) );
+            }
+            else {
+                if (!schedule)
+                    schedule.reset(new Opm::Schedule(*deck, *eclipseState, parseContext, errorGuard, python));
+            }
+            setupMessageLimiter(schedule->getMessageLimits(), "STDOUT_LOGGER");
+            if (!summaryConfig)
+                summaryConfig.reset( new Opm::SummaryConfig(*deck, *schedule, eclipseState->getTableManager(), parseContext, errorGuard));
+#if HAVE_MPI
+            parseSuccess = 1;
+#endif
+        }
+        catch(const std::exception& e)
+        {
+            failureMessage = e.what();
+        }
+    }
+#if HAVE_MPI
+    else {
+        if (!summaryConfig)
+            summaryConfig.reset(new Opm::SummaryConfig);
+        if (!schedule)
+            schedule.reset(new Opm::Schedule(python));
+        if (!eclipseState)
+            eclipseState.reset(new Opm::ParallelEclipseState);
+    }
+
+    auto comm = Dune::MPIHelper::getCollectiveCommunication();
+    parseSuccess = comm.max(parseSuccess);
+    if (!parseSuccess)
+    {
+        if (errorGuard) {
+            errorGuard.dump();
+            errorGuard.clear();
+        }
+        MPI_Abort(MPI_COMM_WORLD, EXIT_FAILURE);
+    }
+
+    Opm::eclStateBroadcast(*eclipseState, *schedule, *summaryConfig);
+#endif
+
+    Opm::checkConsistentArrayDimensions(*eclipseState, *schedule, parseContext, errorGuard);
+
+    if (errorGuard) {
+        errorGuard.dump();
+        errorGuard.clear();
+
+        throw std::runtime_error("Unrecoverable errors were encountered while loading input.");
+    }
+}
+} // end namespace Opm
